Split up the `To()` extension tests
Fixes #60
