--- Tests/Converter.Temperature.Tests/Extensions/To/ToLongExtensionTests.cs
@@ -1,868 +1,0 @@
-﻿namespace Converter.Temperature.Tests.Extensions.To;
-
-using System;
-using FluentAssertions;
-using Temperature.Extensions.To;
-using Temperature.Extensions.To.Celsius;
-using Temperature.Extensions.To.Fahrenheit;
-using Temperature.Extensions.To.Gas;
-using Temperature.Extensions.To.Kelvin;
-using Temperature.Extensions.To.Rankine;
-using TemperatureTypes;
-using Types.Celsius;
-using Types.Fahrenheit;
-using Types.Gas;
-using Types.Kelvin;
-using Types.Rankine;
-using Xunit;
-
-public sealed class ToLongExtensionTests
-{
-    [Fact]
-    public void Test_to_celsius_from_celsius_returns_same_value()
-    {
-        // Arrange.
-        CelsiusLong input = new(42L);
-
-        // Act.
-        long result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_celsius_returns_same_value()
-    {
-        // Arrange.
-        CelsiusLong input = new(42L);
-
-        // Act.
-        long result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_celsius_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 10L;
-        FahrenheitLong input = new(50L);
-
-        // Act.
-        long result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 10L;
-        FahrenheitLong input = new(50L);
-
-        // Act.
-        long result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 230L;
-        GasLong input = new(7L);
-
-        // Act.
-        long result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Theory]
-    [InlineData(1L, 150L)]
-    [InlineData(2L, 165L)]
-    [InlineData(3L, 180L)]
-    [InlineData(4L, 190L)]
-    [InlineData(5L, 202L)]
-    [InlineData(6L, 217L)]
-    [InlineData(7L, 230L)]
-    [InlineData(8L, 240L)]
-    [InlineData(9L, 250L)]
-    public void Test_to_celsius_generic_from_gas_returns_correct_value(
-        long input,
-        long expected)
-    {
-        // Arrange.
-        GasLong inputGas = new(input);
-
-        // Act.
-        long result = inputGas.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 1L;
-        KelvinLong input = new(274L);
-
-        // Act.
-        long result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 1L;
-        KelvinLong input = new(274L);
-
-        // Act.
-        long result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = -121L;
-        RankineLong input = new(274L);
-
-        // Act.
-        long result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = -121L;
-        RankineLong input = new(274L);
-
-        // Act.
-        long result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 54L;
-        CelsiusLong input = new(12L);
-
-        // Act.
-        long result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 54L;
-        CelsiusLong input = new(12L);
-
-        // Act.
-        long result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Theory]
-    [InlineData(long.MinValue)]
-    [InlineData(long.MaxValue)]
-    public void Test_to_fahrenheit_from_celsius_with_invalid_parameter_throws_exception(long input)
-    {
-        // Arrange.
-        CelsiusLong inputCelsius = new(input);
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.ToFahrenheit());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Value out of range for type.");
-    }
-
-    [Theory]
-    [InlineData(long.MinValue)]
-    [InlineData(long.MaxValue)]
-    public void Test_to_fahrenheit_generic_from_celsius_with_invalid_parameter_throws_exception(long input)
-    {
-        // Arrange.
-        CelsiusLong inputCelsius = new(input);
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.To<Fahrenheit>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Value out of range for type.");
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        FahrenheitLong input = new(50L);
-
-        // Act.
-        long result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        FahrenheitLong input = new(50L);
-
-        // Act.
-        long result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 446L;
-        GasLong input = new(7L);
-
-        // Act.
-        long result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 446L;
-        GasLong input = new(7L);
-
-        // Act.
-        long result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 34L;
-        KelvinLong input = new(274L);
-
-        // Act.
-        long result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 34L;
-        KelvinLong input = new(274L);
-
-        // Act.
-        long result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = -186L;
-        RankineLong input = new(274L);
-
-        // Act.
-        long result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = -186L;
-        RankineLong input = new(274L);
-
-        // Act.
-        long result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 6L;
-        CelsiusLong input = new(200L);
-
-        // Act.
-        long result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Theory]
-    [InlineData(140L, 1L)]
-    [InlineData(150L, 2L)]
-    [InlineData(165L, 3L)]
-    [InlineData(180L, 4L)]
-    [InlineData(190L, 5L)]
-    [InlineData(200L, 6L)]
-    [InlineData(220L, 7L)]
-    [InlineData(230L, 8L)]
-    [InlineData(240L, 9L)]
-    public void Test_to_gas_generic_from_celsius_returns_correct_value(
-        long input,
-        long expected)
-    {
-        // Arrange.
-        CelsiusLong inputCelsius = new(input);
-
-        // Act.
-        long result = inputCelsius.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 6L;
-        FahrenheitLong input = new(392L);
-
-        // Act.
-        long result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 6L;
-        FahrenheitLong input = new(392L);
-
-        // Act.
-        long result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_gas_returns_same_value()
-    {
-        // Arrange.
-        GasLong input = new(7L);
-
-        // Act.
-        long result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_gas_returns_same_value()
-    {
-        // Arrange.
-        GasLong input = new(7L);
-
-        // Act.
-        long result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 6L;
-        KelvinLong input = new(473L);
-
-        // Act.
-        long result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 6L;
-        KelvinLong input = new(473L);
-
-        // Act.
-        long result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 6L;
-        RankineLong input = new(862L);
-
-        // Act.
-        long result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 6L;
-        RankineLong input = new(862L);
-
-        // Act.
-        long result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 473L;
-        CelsiusLong input = new(200L);
-
-        // Act.
-        long result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 473L;
-        CelsiusLong input = new(200L);
-
-        // Act.
-        long result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_celsius_with_invalid_parameter_throws_exception()
-    {
-        // Arrange.
-        CelsiusLong inputCelsius = new(long.MaxValue);
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.ToKelvin());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Value out of range for type.");
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_celsius_with_invalid_parameter_throws_exception()
-    {
-        // Arrange.
-        CelsiusLong inputCelsius = new(long.MaxValue);
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.To<Kelvin>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Value out of range for type.");
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 473L;
-        FahrenheitLong input = new(392L);
-
-        // Act.
-        long result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 473L;
-        FahrenheitLong input = new(392L);
-
-        // Act.
-        long result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 490L;
-        GasLong input = new(6L);
-
-        // Act.
-        long result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 490L;
-        GasLong input = new(6L);
-
-        // Act.
-        long result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_kelvin_returns_same_value()
-    {
-        // Arrange.
-        KelvinLong input = new(473L);
-
-        // Act.
-        long result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_kelvin_returns_same_value()
-    {
-        // Arrange.
-        KelvinLong input = new(473L);
-
-        // Act.
-        long result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 478L;
-        RankineLong input = new(862L);
-
-        // Act.
-        long result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 478L;
-        RankineLong input = new(862L);
-
-        // Act.
-        long result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 852L;
-        CelsiusLong input = new(200L);
-
-        // Act.
-        long result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 852L;
-        CelsiusLong input = new(200L);
-
-        // Act.
-        long result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_celsius_with_invalid_parameter_throws_exception()
-    {
-        // Arrange.
-        CelsiusLong inputCelsius = new(long.MaxValue);
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.ToRankine());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Value out of range for type.");
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_celsius_with_invalid_parameter_throws_exception()
-    {
-        // Arrange.
-        CelsiusLong inputCelsius = new(long.MaxValue);
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.To<Rankine>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Value out of range for type.");
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 716L;
-        FahrenheitLong input = new(256L);
-
-        // Act.
-        long result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 716L;
-        FahrenheitLong input = new(256L);
-
-        // Act.
-        long result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 882L;
-        GasLong input = new(6L);
-
-        // Act.
-        long result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 882L;
-        GasLong input = new(6L);
-
-        // Act.
-        long result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 11860L;
-        KelvinLong input = new(6589L);
-
-        // Act.
-        long result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const long expected = 11860L;
-        KelvinLong input = new(6589L);
-
-        // Act.
-        long result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_rankine_returns_same_value()
-    {
-        // Arrange.
-        RankineLong input = new(862L);
-
-        // Act.
-        long result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_rankine_returns_same_value()
-    {
-        // Arrange.
-        RankineLong input = new(862L);
-
-        // Act.
-        long result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-}
