--- Tests/Converter.Temperature.Tests/Extensions/To/ToStringExtensionsTests.cs
@@ -1,1017 +1,0 @@
-﻿namespace Converter.Temperature.Tests.Extensions.To;
-
-using System;
-using System.Globalization;
-using FluentAssertions;
-using Temperature.Extensions.To;
-using Temperature.Extensions.To.Celsius;
-using Temperature.Extensions.To.Fahrenheit;
-using Temperature.Extensions.To.Gas;
-using Temperature.Extensions.To.Kelvin;
-using Temperature.Extensions.To.Rankine;
-using TemperatureTypes;
-using Types.Celsius;
-using Types.Fahrenheit;
-using Types.Gas;
-using Types.Kelvin;
-using Types.Rankine;
-using Xunit;
-
-public sealed class ToStringExtensionsTests
-{
-    [Fact]
-    public void Test_to_celsius_from_celsius_returns_same_value()
-    {
-        // Arrange.
-        CelsiusString input = new("42");
-
-        // Act.
-        string result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_celsius_returns_same_value()
-    {
-        // Arrange.
-        CelsiusString input = new("42");
-
-        // Act.
-        string result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_celsius_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "10";
-        FahrenheitString input = new("50");
-
-        // Act.
-        string result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "10";
-        FahrenheitString input = new("50");
-
-        // Act.
-        string result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "220";
-        GasString input = new("7");
-
-        // Act.
-        string result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "220";
-        GasString input = new("7");
-
-        // Act.
-        string result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "1";
-        KelvinString input = new("274.15");
-
-        // Act.
-        string result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "1";
-        KelvinString input = new("274.15");
-
-        // Act.
-        string result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "1.0000000000000062";
-        RankineString input = new("493.47");
-
-        // Act.
-        string result = input.ToCelsius();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_celsius_generic_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "1.0000000000000062";
-        RankineString input = new("493.47");
-
-        // Act.
-        string result = input.To<Celsius>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_celsius_returns_same_value()
-    {
-        // Arrange.
-        const string expected = "53.6";
-        CelsiusString input = new("12");
-
-        // Act.
-        string result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_celsius_returns_same_value()
-    {
-        // Arrange.
-        const string expected = "53.6";
-        CelsiusString input = new("12");
-
-        // Act.
-        string result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        FahrenheitString input = new("50");
-
-        // Act.
-        string result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        FahrenheitString input = new("50");
-
-        // Act.
-        string result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "428";
-        GasString input = new("7");
-
-        // Act.
-        string result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "428";
-        GasString input = new("7");
-
-        // Act.
-        string result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "33.8";
-        KelvinString input = new("274.15");
-
-        // Act.
-        string result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "33.8";
-        KelvinString input = new("274.15");
-
-        // Act.
-        string result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "33.80099999999999";
-        RankineString input = new("493.471");
-
-        // Act.
-        string result = input.ToFahrenheit();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_fahrenheit_generic_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "33.80099999999999";
-        RankineString input = new("493.471");
-
-        // Act.
-        string result = input.To<Fahrenheit>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "6";
-        CelsiusString input = new("200");
-
-        // Act.
-        string result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "6";
-        CelsiusString input = new("200");
-
-        // Act.
-        string result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_celsius_throws_out_of_range_exception_too_low()
-    {
-        // Arrange.
-        CelsiusString input = new("79");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.ToGas());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too low for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_celsius_throws_out_of_range_exception_too_low()
-    {
-        // Arrange.
-        CelsiusString input = new("79");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.To<Gas>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too low for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_from_celsius_throws_out_of_range_exception_too_high()
-    {
-        // Arrange.
-        CelsiusString input = new("271");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.ToGas());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too high for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_celsius_throws_out_of_range_exception_too_high()
-    {
-        // Arrange.
-        CelsiusString input = new("271");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.To<Gas>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too high for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "6";
-        FahrenheitString input = new("392");
-
-        // Act.
-        string result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "6";
-        FahrenheitString input = new("392");
-
-        // Act.
-        string result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_fahrenheit_throws_out_of_range_exception_too_low()
-    {
-        // Arrange.
-        FahrenheitString input = new("174");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.ToGas());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too low for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_fahrenheit_throws_out_of_range_exception_too_low()
-    {
-        // Arrange.
-        FahrenheitString input = new("174");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.To<Gas>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too low for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_from_fahrenheit_throws_out_of_range_exception_too_high()
-    {
-        // Arrange.
-        FahrenheitString input = new("520");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.ToGas());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too high for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_fahrenheit_throws_out_of_range_exception_too_high()
-    {
-        // Arrange.
-        FahrenheitString input = new("520");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.To<Gas>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too high for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_from_gas_returns_same_value()
-    {
-        // Arrange.
-        GasString input = new("7");
-
-        // Act.
-        string result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_gas_returns_same_value()
-    {
-        // Arrange.
-        GasString input = new("7");
-
-        // Act.
-        string result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "6";
-        KelvinString input = new("478.15");
-
-        // Act.
-        string result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_kelvin_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "6";
-        KelvinString input = new("478.15");
-
-        // Act.
-        string result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_kelvin_throws_out_of_range_exception_too_low()
-    {
-        // Arrange.
-        KelvinString input = new("174");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.ToGas());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too low for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_kelvin_throws_out_of_range_exception_too_low()
-    {
-        // Arrange.
-        KelvinString input = new("174");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.To<Gas>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too low for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_from_kelvin_throws_out_of_range_exception_too_high()
-    {
-        // Arrange.
-        KelvinString input = new("544.15");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.ToGas());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too high for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_kelvin_throws_out_of_range_exception_too_high()
-    {
-        // Arrange.
-        KelvinString input = new("544.15");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.To<Gas>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too high for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "3";
-        RankineString input = new("806.67");
-
-        // Act.
-        string result = input.ToGas();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "3";
-        RankineString input = new("806.67");
-
-        // Act.
-        string result = input.To<Gas>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_gas_from_rankine_throws_out_of_range_exception_too_low()
-    {
-        // Arrange.
-        RankineString input = new("174");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.ToGas());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too low for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_rankine_throws_out_of_range_exception_too_low()
-    {
-        // Arrange.
-        RankineString input = new("174");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.To<Gas>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too low for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_from_rankine_throws_out_of_range_exception_too_high()
-    {
-        // Arrange.
-        RankineString input = new("1524.25");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.ToGas());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too high for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_gas_generic_from_rankine_throws_out_of_range_exception_too_high()
-    {
-        // Arrange.
-        RankineString input = new("1524.25");
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => input.To<Gas>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Temp too high for gas mark!");
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "473.15";
-        CelsiusString input = new("200");
-
-        // Act.
-        string result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "473.15";
-        CelsiusString input = new("200");
-
-        // Act.
-        string result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "473.15";
-        FahrenheitString input = new("392");
-
-        // Act.
-        string result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "473.15";
-        FahrenheitString input = new("392");
-
-        // Act.
-        string result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "473.15";
-        GasString input = new("6");
-
-        // Act.
-        string result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "473.15";
-        GasString input = new("6");
-
-        // Act.
-        string result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_kelvin_returns_same_value()
-    {
-        // Arrange.
-        KelvinString input = new("473.15");
-
-        // Act.
-        string result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_kelvin_returns_same_value()
-    {
-        // Arrange.
-        KelvinString input = new("473.15");
-
-        // Act.
-        string result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "473.1499999999999";
-        RankineString input = new("851.67");
-
-        // Act.
-        string result = input.ToKelvin();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_kelvin_generic_from_rankine_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "473.1499999999999";
-        RankineString input = new("851.67");
-
-        // Act.
-        string result = input.To<Kelvin>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "851.6699999999998";
-        CelsiusString input = new("200");
-
-        // Act.
-        string result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_celsius_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "851.6699999999998";
-        CelsiusString input = new("200");
-
-        // Act.
-        string result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Theory]
-    [InlineData(double.MinValue)]
-    [InlineData(double.MaxValue)]
-    public void Test_to_rankine_from_celsius_with_invalid_parameter_throws_exception(double input)
-    {
-        // Arrange.
-        CelsiusString inputCelsius = new(input.ToString(CultureInfo.InvariantCulture));
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.ToRankine());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Value out of range for type.");
-    }
-
-    [Theory]
-    [InlineData(double.MinValue)]
-    [InlineData(double.MaxValue)]
-    public void Test_to_rankine_generic_from_celsius_with_invalid_parameter_throws_exception(double input)
-    {
-        // Arrange.
-        CelsiusString inputCelsius = new(input.ToString(CultureInfo.InvariantCulture));
-
-        // Act.
-        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.To<Rankine>());
-
-        // Assert.
-        result.Message.Should()
-            .Contain("Value out of range for type.");
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "851.6700000000001";
-        FahrenheitString input = new("392");
-
-        // Act.
-        string result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_fahrenheit_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "851.6700000000001";
-        FahrenheitString input = new("392");
-
-        // Act.
-        string result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "851.6699999999998";
-        GasString input = new("6");
-
-        // Act.
-        string result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_gas_returns_correct_value()
-    {
-        // Arrange.
-        const string expected = "851.6699999999998";
-        GasString input = new("6");
-
-        // Act.
-        string result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_kelvin_returns_same_value()
-    {
-        // Arrange.
-        const string expected = "851.6699999999998";
-        KelvinString input = new("473.15");
-
-        // Act.
-        string result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_kelvin_returns_same_value()
-    {
-        // Arrange.
-        const string expected = "851.6699999999998";
-        KelvinString input = new("473.15");
-
-        // Act.
-        string result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(expected);
-    }
-
-    [Fact]
-    public void Test_to_rankine_from_rankine_returns_same_value()
-    {
-        // Arrange.
-        RankineString input = new("851.67");
-
-        // Act.
-        string result = input.ToRankine();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-
-    [Fact]
-    public void Test_to_rankine_generic_from_rankine_returns_same_value()
-    {
-        // Arrange.
-        RankineString input = new("851.67");
-
-        // Act.
-        string result = input.To<Rankine>();
-
-        // Assert.
-        result.Should()
-            .Be(input.Temperature);
-    }
-}
