+++ tests/Converter.Temperature.Tests/Extensions/To/Kelvin/ToFloatExtensionTests.cs
@@ -1,0 +1,163 @@
+﻿namespace Converter.Temperature.Tests.Extensions.To.Kelvin;
+
+using FluentAssertions;
+using Temperature.Extensions.To;
+using Temperature.Extensions.To.Kelvin;
+using TemperatureTypes;
+using Types.Celsius;
+using Types.Fahrenheit;
+using Types.Gas;
+using Types.Kelvin;
+using Types.Rankine;
+using Xunit;
+
+public sealed class ToFloatExtensionTests
+{
+    [Fact]
+    public void Test_to_kelvin_from_celsius_returns_correct_value()
+    {
+        // Arrange.
+        const float expected = 473.15f;
+        CelsiusFloat input = new(200);
+
+        // Act.
+        float result = input.ToKelvin();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_generic_from_celsius_returns_correct_value()
+    {
+        // Arrange.
+        const float expected = 473.15f;
+        CelsiusFloat input = new(200);
+
+        // Act.
+        float result = input.To<Kelvin>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_from_fahrenheit_returns_correct_value()
+    {
+        // Arrange.
+        const float expected = 473.15f;
+        FahrenheitFloat input = new(392);
+
+        // Act.
+        float result = input.ToKelvin();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_generic_from_fahrenheit_returns_correct_value()
+    {
+        // Arrange.
+        const float expected = 473.15f;
+        FahrenheitFloat input = new(392);
+
+        // Act.
+        float result = input.To<Kelvin>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_from_gas_returns_correct_value()
+    {
+        // Arrange.
+        const float expected = 473.15f;
+        GasFloat input = new(6);
+
+        // Act.
+        float result = input.ToKelvin();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_generic_from_gas_returns_correct_value()
+    {
+        // Arrange.
+        const float expected = 473.15f;
+        GasFloat input = new(6);
+
+        // Act.
+        float result = input.To<Kelvin>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_from_kelvin_returns_same_value()
+    {
+        // Arrange.
+        KelvinFloat input = new(473.15f);
+
+        // Act.
+        float result = input.ToKelvin();
+
+        // Assert.
+        result.Should()
+            .Be(input.Temperature);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_generic_from_kelvin_returns_same_value()
+    {
+        // Arrange.
+        KelvinFloat input = new(473.15f);
+
+        // Act.
+        float result = input.To<Kelvin>();
+
+        // Assert.
+        result.Should()
+            .Be(input.Temperature);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_from_rankine_returns_correct_value()
+    {
+        // Arrange.
+        const float expected = 473.15002f;
+        RankineFloat input = new(851.67f);
+
+        // Act.
+        float result = input.ToKelvin();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_kelvin_generic_from_rankine_returns_correct_value()
+    {
+        // Arrange.
+        const float expected = 473.15002f;
+        RankineFloat input = new(851.67f);
+
+        // Act.
+        float result = input.To<Kelvin>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+}
