+++ tests/Converter.Temperature.Tests/Extensions/To/Rankine/ToDoubleExtensionTests.cs
@@ -1,0 +1,197 @@
+﻿namespace Converter.Temperature.Tests.Extensions.To.Rankine;
+
+using System;
+using FluentAssertions;
+using Temperature.Extensions.To;
+using Temperature.Extensions.To.Rankine;
+using TemperatureTypes;
+using Types.Celsius;
+using Types.Fahrenheit;
+using Types.Gas;
+using Types.Kelvin;
+using Types.Rankine;
+using Xunit;
+
+public sealed class ToDoubleExtensionTests
+{
+    [Fact]
+    public void Test_to_rankine_from_celsius_returns_correct_value()
+    {
+        // Arrange.
+        const double expected = 851.6699999999998d;
+        CelsiusDouble input = new(200);
+
+        // Act.
+        double result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_celsius_returns_correct_value()
+    {
+        // Arrange.
+        const double expected = 851.6699999999998d;
+        CelsiusDouble input = new(200);
+
+        // Act.
+        double result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Theory]
+    [InlineData(double.MinValue)]
+    [InlineData(double.MaxValue)]
+    public void Test_to_rankine_from_celsius_with_invalid_parameter_throws_exception(double input)
+    {
+        // Arrange.
+        CelsiusDouble inputCelsius = new(input);
+
+        // Act.
+        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.ToRankine());
+
+        // Assert.
+        result.Message.Should()
+            .Contain("Value out of range for type.");
+    }
+
+    [Theory]
+    [InlineData(double.MinValue)]
+    [InlineData(double.MaxValue)]
+    public void Test_to_rankine_generic_from_celsius_with_invalid_parameter_throws_exception(double input)
+    {
+        // Arrange.
+        CelsiusDouble inputCelsius = new(input);
+
+        // Act.
+        ArgumentOutOfRangeException result =
+            Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.To<Rankine>());
+
+        // Assert.
+        result.Message.Should()
+            .Contain("Value out of range for type.");
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_fahrenheit_returns_correct_value()
+    {
+        // Arrange.
+        const double expected = 851.6700000000001d;
+        FahrenheitDouble input = new(392);
+
+        // Act.
+        double result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_fahrenheit_returns_correct_value()
+    {
+        // Arrange.
+        const double expected = 851.6700000000001d;
+        FahrenheitDouble input = new(392);
+
+        // Act.
+        double result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_gas_returns_correct_value()
+    {
+        // Arrange.
+        const double expected = 851.6699999999998d;
+        GasDouble input = new(6);
+
+        // Act.
+        double result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_gas_returns_correct_value()
+    {
+        // Arrange.
+        const double expected = 851.6699999999998d;
+        GasDouble input = new(6);
+
+        // Act.
+        double result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_kelvin_returns_same_value()
+    {
+        // Arrange.
+        const double expected = 851.6699999999998d;
+        KelvinDouble input = new(473.15);
+
+        // Act.
+        double result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_kelvin_returns_same_value()
+    {
+        // Arrange.
+        const double expected = 851.6699999999998d;
+        KelvinDouble input = new(473.15);
+
+        // Act.
+        double result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_rankine_returns_same_value()
+    {
+        // Arrange.
+        RankineDouble input = new(851.67);
+
+        // Act.
+        double result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(input.Temperature);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_rankine_returns_same_value()
+    {
+        // Arrange.
+        RankineDouble input = new(851.67);
+
+        // Act.
+        double result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(input.Temperature);
+    }
+}
