+++ tests/Converter.Temperature.Tests/Extensions/To/Rankine/ToLongExtensionTests.cs
@@ -1,0 +1,192 @@
+﻿namespace Converter.Temperature.Tests.Extensions.To.Rankine;
+
+using System;
+using FluentAssertions;
+using Temperature.Extensions.To;
+using Temperature.Extensions.To.Rankine;
+using TemperatureTypes;
+using Types.Celsius;
+using Types.Fahrenheit;
+using Types.Gas;
+using Types.Kelvin;
+using Types.Rankine;
+using Xunit;
+
+public sealed class ToLongExtensionTests
+{
+    [Fact]
+    public void Test_to_rankine_from_celsius_returns_correct_value()
+    {
+        // Arrange.
+        const long expected = 852L;
+        CelsiusLong input = new(200L);
+
+        // Act.
+        long result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_celsius_returns_correct_value()
+    {
+        // Arrange.
+        const long expected = 852L;
+        CelsiusLong input = new(200L);
+
+        // Act.
+        long result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_celsius_with_invalid_parameter_throws_exception()
+    {
+        // Arrange.
+        CelsiusLong inputCelsius = new(long.MaxValue);
+
+        // Act.
+        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.ToRankine());
+
+        // Assert.
+        result.Message.Should()
+            .Contain("Value out of range for type.");
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_celsius_with_invalid_parameter_throws_exception()
+    {
+        // Arrange.
+        CelsiusLong inputCelsius = new(long.MaxValue);
+
+        // Act.
+        ArgumentOutOfRangeException result = Assert.Throws<ArgumentOutOfRangeException>(() => inputCelsius.To<Rankine>());
+
+        // Assert.
+        result.Message.Should()
+            .Contain("Value out of range for type.");
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_fahrenheit_returns_correct_value()
+    {
+        // Arrange.
+        const long expected = 716L;
+        FahrenheitLong input = new(256L);
+
+        // Act.
+        long result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_fahrenheit_returns_correct_value()
+    {
+        // Arrange.
+        const long expected = 716L;
+        FahrenheitLong input = new(256L);
+
+        // Act.
+        long result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_gas_returns_correct_value()
+    {
+        // Arrange.
+        const long expected = 882L;
+        GasLong input = new(6L);
+
+        // Act.
+        long result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_gas_returns_correct_value()
+    {
+        // Arrange.
+        const long expected = 882L;
+        GasLong input = new(6L);
+
+        // Act.
+        long result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_kelvin_returns_correct_value()
+    {
+        // Arrange.
+        const long expected = 11860L;
+        KelvinLong input = new(6589L);
+
+        // Act.
+        long result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_kelvin_returns_correct_value()
+    {
+        // Arrange.
+        const long expected = 11860L;
+        KelvinLong input = new(6589L);
+
+        // Act.
+        long result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(expected);
+    }
+
+    [Fact]
+    public void Test_to_rankine_from_rankine_returns_same_value()
+    {
+        // Arrange.
+        RankineLong input = new(862L);
+
+        // Act.
+        long result = input.ToRankine();
+
+        // Assert.
+        result.Should()
+            .Be(input.Temperature);
+    }
+
+    [Fact]
+    public void Test_to_rankine_generic_from_rankine_returns_same_value()
+    {
+        // Arrange.
+        RankineLong input = new(862L);
+
+        // Act.
+        long result = input.To<Rankine>();
+
+        // Assert.
+        result.Should()
+            .Be(input.Temperature);
+    }
+}
